Update file name and refactor changes.
