+++ benchmarks/benchmark_restoration.py
@@ -156,7 +156,7 @@ class RollingBall:
     time_rollingball_nan.param_names = ["radius"]
 
     def time_rollingball_ndim(self):
-        from skimage.restoration.rolling_ball import ellipsoid_kernel
+        from skimage.restoration._rolling_ball import ellipsoid_kernel
         image = data.cells3d()[:, 1, ...]
         kernel = ellipsoid_kernel((1, 100, 100), 100)
         restoration.rolling_ball(
