+++ skimage/restoration/_rolling_ball.py
